+++ test/endpoint.js
@@ -12,7 +12,6 @@ if (util.inspect.defaultOptions)
 describe('helpers', function() {
   require('./helpers/index.js');
   require('./helpers/matrices.js');
-  require('./helpers/random.js');
   require('./helpers/vectors.js');
 });